pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.124.2",
    "openai>=2.11.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "bcrypt>=4.2.0",
    "google-generativeai>=0.8.6",
    "llama-parse>=0.6.92",
]
